Adding validation for server

+++ src/main/java/com/example/taskserver/controller/TaskController.java
@@ -28,6 +28,9 @@ public class TaskController {
 
     @PostMapping
     public ResponseEntity<Task> saveTask(@Valid @RequestBody Task task, Errors errors, HttpServletRequest request) {
+        if (!service.isTokenValidBoss(request)) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         log.info("Task for saving: {}", task);
         if (errors.hasErrors()) {
             log.error("Not valid task: {}", task);
@@ -43,8 +46,8 @@ public class TaskController {
 
     @GetMapping
     public ResponseEntity<?> getTask(@RequestParam(value = "taskName", required = false) String name, HttpServletRequest request) {
-        if (!service.isTokenValid(request)) {
-    throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        if (!service.isTokenValidBoss(request)) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
         }
         if (name == null) {
             log.info("Getting all tasks");
@@ -58,6 +61,9 @@ public class TaskController {
 
     @GetMapping("/{id}")
     public ResponseEntity<Task> findTaskById(@PathVariable("id") Long id, HttpServletRequest request) {
+        if (!service.isTokenValidBossAndUser(request)) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         log.info("Searching task with id: {}", id);
         Task task = service.findById(id);
         if (Objects.isNull(task)) {
@@ -70,26 +76,35 @@ public class TaskController {
 
 
     @DeleteMapping("/{id}")
-    public void deleteTask(@PathVariable("id") Long id) {
+    public void deleteTask(@PathVariable("id") Long id ,HttpServletRequest request) {
+        if (!service.isTokenValidBoss(request)) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         log.info("Deleting task with id: {}", id);
         service.delete(id);
     }
 
     @PatchMapping("/{id}")
     public ResponseEntity<Task> updateTask(@PathVariable("id") Long id, @RequestBody Task task, HttpServletRequest request) {
+        if (!service.isTokenValidBoss(request)) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         log.info("Task with id: {}. And body: {}", id, task);
         return ResponseEntity.ok(service.update(id, task));
     }
 
     @PatchMapping("/{id}/addBand")
     public ResponseEntity<Task> addToBand(@PathVariable("id") Long id, @RequestBody Band bandName, HttpServletRequest request) {
+        if (!service.isTokenValidBoss(request)) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         log.info("Add task with name {}", bandName);
         service.addTaskToBand(id, bandName);
         return ResponseEntity.ok(service.findById(id));
     }
 
     @PatchMapping("/{id}/completed")
-    public ResponseEntity<Task> makeCompleted(@PathVariable("id") Long id, HttpServletRequest request) {
+    public ResponseEntity<Task> makeCompleted(@PathVariable("id") Long id) {
         log.info("Make task with id: {} completed", id);
         service.makeTaskCompleted(id);
         return ResponseEntity.ok(service.findById(id));
